legacy/requirements.txt
discord[voice]
pynacl
youtube-dl
python-dotenv